+++ mmms-backend/src/main/java/com/sep/mmms_backend/controller/RegisterUserController.java
@@ -1,8 +1,9 @@
 package com.sep.mmms_backend.controller;
 
 import com.sep.mmms_backend.entity.AppUser;
+import com.sep.mmms_backend.exceptions.UsernameAlreadyExistsException;
 import com.sep.mmms_backend.response.Response;
-import com.sep.mmms_backend.response.ResponseMessage;
+import com.sep.mmms_backend.response.ResponseMessages;
 import com.sep.mmms_backend.service.AppUserService;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
@@ -13,7 +14,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import java.util.ArrayList;
+import java.util.HashMap;
 
 @RestController
 @Slf4j
@@ -30,17 +32,45 @@ public class RegisterUserController {
     @PostMapping("/register")
     public ResponseEntity<Response> registerUser(@RequestBody @Valid AppUser appUser, Errors errors) {
         if(errors.hasErrors()){
-            List<String> errorMessages = errors.getFieldErrors()
-                    .stream()
-                    .map(error -> error.getField() + ": " + error.getDefaultMessage())
-                    .toList();
+           HashMap<String, ArrayList<String>> errorMessages = new HashMap<>();
+
+//           errors.getFieldErrors()
+//                    .forEach(error-> errorMessages.put(error.getField(), error.getDefaultMessage()));
+
+            errors.getFieldErrors().forEach(
+                   error-> {
+                       //if the key is not already present in the Map, create the key as well as new ArrayList for the value
+                       if(!errorMessages.containsKey(error.getField())){
+                           ArrayList<String> list = new ArrayList<>();
+                           list.add(error.getDefaultMessage());
+                           errorMessages.put(error.getField(),list);
+                       } else {
+                           errorMessages.get(error.getField()).add(error.getDefaultMessage());
+                       }
+                   }
+            );
             log.error("Validation Failed while registering user: {}", errorMessages);
-            return ResponseEntity.ok().body(new Response(ResponseMessage.REGISTER_USER_FAILED, errorMessages));
+
+
+            return ResponseEntity.badRequest().body(new Response(ResponseMessages.REGISTER_USER_FAILED, errorMessages));
         }
 
         appUser.setPassword(passwordEncoder.encode(appUser.getPassword()));
-        appUserService.saveUser(appUser);
+        //save the data to the database
+        AppUser savedUser = null;
+
+        //validate that the username is unique
+        try {
+            savedUser = appUserService.saveNewUser(appUser);
+        } catch(UsernameAlreadyExistsException e) {
+            log.error("User with the username `{}` already exists", appUser.getUsername());
+            return ResponseEntity.badRequest().body(new Response(e.getMessage()));
+        }
+        if(savedUser != null && savedUser.getUid() <=0) {
+            return ResponseEntity.internalServerError().body(null);
+        }
+
         log.info("User with the username @{} registered successfully", appUser.getUsername());
-        return ResponseEntity.ok().body(new Response(ResponseMessage.REGISTER_USER_SUCCESS));
+        return ResponseEntity.ok().body(new Response(ResponseMessages.REGISTER_USER_SUCCESS));
     }
 }