+++ mmms-backend/src/main/java/com/sep/mmms_backend/exceptions/ExceptionMessages.java
@@ -1,0 +1,16 @@
+package com.sep.mmms_backend.exceptions;
+
+
+public enum ExceptionMessages {
+    USERNAME_ALREADY_EXISTS("user with the username already exists"),
+    ;
+
+    final private String message;
+    ExceptionMessages(String message) {
+        this.message = message;
+    }
+
+    public String toString() {
+        return message;
+    }
+}