+++ mmms-backend/src/main/java/com/sep/mmms_backend/exceptions/UsernameAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.sep.mmms_backend.exceptions;
+
+public class UsernameAlreadyExistsException extends RuntimeException {
+    public UsernameAlreadyExistsException(String message) {
+        super(message);
+    }
+}