+++ mmms-backend/src/main/java/com/sep/mmms_backend/repository/AppUserRepository.java
@@ -9,4 +9,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
     public AppUser findByUsername(String username);
+    public boolean existsByUsername(String username);
 }