+++ mmms-backend/src/main/java/com/sep/mmms_backend/service/AppUserService.java
@@ -1,6 +1,8 @@
 package com.sep.mmms_backend.service;
 
 import com.sep.mmms_backend.entity.AppUser;
+import com.sep.mmms_backend.exceptions.ExceptionMessages;
+import com.sep.mmms_backend.exceptions.UsernameAlreadyExistsException;
 import com.sep.mmms_backend.repository.AppUserRepository;
 import org.springframework.stereotype.Service;
 
@@ -17,7 +19,14 @@ public class AppUserService {
         return appUserRepository.findByUsername(username);
     }
 
-    public AppUser saveUser(AppUser appUser) {
+    public AppUser saveNewUser(AppUser appUser) {
+
+        if(appUserRepository.existsByUsername(appUser.getUsername())) {
+            throw new UsernameAlreadyExistsException(ExceptionMessages.USERNAME_ALREADY_EXISTS.toString());
+        }
+
         return appUserRepository.save(appUser);
     }
+
+
 }