+++ mmms-backend/src/main/java/com/sep/mmms_backend/validators/annotations/UsernameFormat.java
@@ -1,6 +1,6 @@
 package com.sep.mmms_backend.validators.annotations;
 
-import com.sep.mmms_backend.validators.CheckUsernameAvailabilityValidator;
+import com.sep.mmms_backend.validators.UsernameFormatValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -9,14 +9,12 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Constraint(validatedBy = CheckUsernameAvailabilityValidator.class)
+@Constraint(validatedBy = UsernameFormatValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
 
-public @interface CheckUsernameAvailability {
-    String message() default "username unavailable";
-
+public @interface UsernameFormat {
+    String message() default "Invalid username format";
     Class<?>[] groups() default {};
-
     Class<? extends Payload>[] payload() default {};
 }